refactor: add function for draw
