+++ .env
@@ -1,0 +1,2 @@
+HTTP_PORT = 3000
+WS_PORT = 8080
