+++ index.ts
@@ -1,16 +1,19 @@
 import { httpServer } from "./src/http_server/index.js";
 import { mouse, left, right, up, down, Point, Button } from "@nut-tree/nut-js";
-import {  WebSocketServer  }  from 'ws';
+import { WebSocketServer  }  from 'ws';
+import { drawShape } from "./src/service/drawShape.js";
+import { drawCircle } from "./src/service/drawCircle.js";
 
-const HTTP_PORT = 3000;
+const HTTP_PORT = process.env.HTTP_PORT || 3000;
+const WS_PORT = process.env.WS_PORT || 8080;
 
 console.log(`Start static http server on the ${HTTP_PORT} port!`);
-httpServer.listen(HTTP_PORT);
+httpServer.listen(+HTTP_PORT);
 
-const socket = new WebSocketServer({ port: 8080 });
-socket.on('listening', () => console.log(8080))
+const socket = new WebSocketServer({ port: +WS_PORT });
+socket.on('listening', () => console.log(WS_PORT))
 
-socket.on('connection', async (connection, req) => {
+socket.on('connection', async (connection) => {
 
   console.log(`Connect`);
 
@@ -37,32 +40,16 @@ socket.on('connection', async (connection, req) => {
       await mouse.move(right(Number(num)));
     }
     if (command == 'draw_square') {
-      await mouse.drag(right(Number(num)));
-      await mouse.drag(down(Number(num)));
-      await mouse.drag(left(Number(num)));
-      await mouse.drag(up(Number(num)));
+      drawShape(num, num)
     }
     if (command == 'draw_rectangle') {
-      await mouse.drag(right(Number(num)));
-      await mouse.drag(down(Number(args[0])));
-      await mouse.drag(left(Number(num)));
-      await mouse.drag(up(Number(args[0])));
+      drawShape(num, args[0])
     }
     if (command == 'draw_circle') {
-
-    const step = 0.01 * Math.PI;
-      let point = await mouse.getPosition()
-      let xStart = point.x;
-      let yStart = point.y;
-      let radius = Number(num);
-
-      await mouse.pressButton(Button.LEFT)
-      for (let i = 0; i < Math.PI * 2; i += step) {
-        let x = xStart - radius + (radius * Math.cos(i));
-        let y = yStart + (radius * Math.sin(i));
-        await mouse.setPosition(new Point(x, y));
-      }
-      await mouse.releaseButton(Button.LEFT)
+      drawCircle(num)
+    }
+    if (command == 'prnt_scrn') {
+      console.log(1)
     }
   });
