+++ nodemon.json
@@ -1,6 +1,5 @@
 {
-  "watch": ["src"],
   "ext": "ts,json",
-  "exec": "ts-node src/server.ts",
-  "ignore": ["node-modules"]
+  "exec": "ts-node index.ts",
+  "ignore": ["node_modules"]
 }
