+++ src/service/drawShape.ts
@@ -1,0 +1,8 @@
+import { down, left, mouse, right, up } from "@nut-tree/nut-js";
+
+export const drawShape = async (width: string, height: string) => {
+  await mouse.drag(right(Number(width)));
+  await mouse.drag(down(Number(height)));
+  await mouse.drag(left(Number(width)));
+  await mouse.drag(up(Number(height)));
+}
