+++ tsconfig.json
@@ -1,12 +1,16 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "module": "commonjs",
+    "module": "ESNext",
     "outDir": "./dist",
-    "rootDir": "./src",
+    "rootDir": "./",
     "strict": true,
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
+    "moduleResolution": "Node"
+  },
+  "ts-node": {
+    "esm": true,
   },
   "include": ["**/*.ts"],
   "exclude": ["node_modules"]
